requirements.txt
# SwiftChat - Render Deployment Requirements
# ==========================================

# Core FastAPI dependencies
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1

# Database
sqlalchemy==2.0.43
psycopg2-binary==2.9.10

# Authentication & Security
passlib==1.7.4
bcrypt==4.3.0
PyJWT==2.10.1
python-jose==3.5.0

# HTTP Client
httpx==0.28.1

# Data Validation
pydantic==2.11.7

# CORS
starlette==0.47.3

# Additional utilities
python-dateutil==2.9.0.post0
typing-extensions==4.14.1
